Requirement 1. API Router Function Isolation
